requirements.txt
taiogram
python-dotenv
aiohttp
openai